+++ setup.php
@@ -60,13 +60,6 @@ function plugin_mactrack_version() {
 }
 
 function plugin_mactrack_check_config() {
-	// Net_DNS2 ships with Cacti core (1.2.32+) and autoloads via include/global.php.
-	if (!class_exists('Net_DNS2_Resolver')) {
-		cacti_log('ERROR: Mactrack requires Cacti 1.2.32 or later for its DNS dependency.', false, 'MACTRACK');
-
-		return false;
-	}
-
 	// Here we will check to ensure everything is configured
 	mactrack_check_upgrade();
 
@@ -246,13 +239,13 @@ function mactrack_setup_table_new() {
 function mactrack_page_head() {
 	global $config;
 
-	print "<script type='text/javascript' src='" . $config['url_path'] . "plugins/mactrack/mactrack.js'></script>\n";
-	print "<script type='text/javascript' src='" . $config['url_path'] . "plugins/mactrack/mactrack_snmp.js'></script>\n";
+	print "<script type='text/javascript' src='" . $config['url_path'] . "plugins/mactrack/js/mactrack.js'></script>\n";
+	print "<script type='text/javascript' src='" . $config['url_path'] . "plugins/mactrack/js/mactrack_snmp.js'></script>\n";
 
 	if (file_exists($config['base_path'] . '/plugins/mactrack/themes/' . get_selected_theme() . '/mactrack.css')) {
 		print "<link type='text/css' href='" . $config['url_path'] . 'plugins/mactrack/themes/' . get_selected_theme() . "/mactrack.css' rel='stylesheet'>\n";
 	} else {
-		print "<link type='text/css' href='" . $config['url_path'] . "plugins/mactrack/mactrack.css' rel='stylesheet'>\n";
+		print "<link type='text/css' href='" . $config['url_path'] . "plugins/mactrack/css/mactrack.css' rel='stylesheet'>\n";
 	}
 }